Adiciona exercicio 7

+++ src/main/java/com/example/exercicio5/models/Cliente.java
@@ -10,5 +10,5 @@ public class Cliente {
     int id;
     String nome;
     @CPF
-    int cpf;
+    long cpf;
 }
